feat : entity 수정 #7
